Move checkout initialization to subscription management page
